Json update checking & app logging.
Wired up the json version checking classes to the menu and startup update checks.
Removed the app logging menu option and associated logic. App logging to the fiddler log is now on without an option to turn it off. Cannot see a reason why this would be an issue, all my testing thus far has been with the logging turned on. Can't think of any good reason to turn off the logging info.